feat: scalene: partially support unbuffered tracing output
You need to use the `-u` option with Python for it to print to
the output window in real time. Otherwise, it buffers and treats the
output as a pipe. Unfortunately, this flag is not working with Scalene,
so you only get to see the buffered output that loads when your
program finishes.

Suggested temporary workaround: Flush your prints with
`sys.stdout.flush()`